+++ packages/react-misskey/src/components/mfm_scale.tsx
@@ -10,13 +10,13 @@ export const MfmScale: FC<Props> = ({ node }) => {
     node.props.args.x === true
       ? 1
       : typeof node.props.args.x === "string"
-      ? parseInt(node.props.args.x)
+      ? parseFloat(node.props.args.x)
       : 1;
   const y =
     node.props.args.y === true
       ? 1
       : typeof node.props.args.y === "string"
-      ? parseInt(node.props.args.y)
+      ? parseFloat(node.props.args.y)
       : 1;
 
   return (
